requirements.txt
Django==4.0.3
Djangorestframework==3.13.1
Pillow==9.0.1
psycopg2==2.9.3
djangorestframework-filters==0.11.1
django-rest-swagger==2.2.0
drf-yasg==1.20.0

tzdata==2021.5
whitenoise==6.0.0
gunicorn==20.1.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
# psycopg2==2.9.3
django-cors-headers==3.11.0
cloudinary==1.29.0
django-cloudinary-storage==0.3.0
python-decouple==3.6
# Pillow==9.0.1

PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
# tzdata==2021.5
djangorestframework-simplejwt==5.1.0
asgiref==3.5.0
numpy==1.22.3
django-filter==1.1
geopy==2.2.0
# drf-yasg==1.20.0
joblib==1.1.0